put dashboard into a Blueprint
