+++ app/auth/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint("auth", __name__)
 
-from app.auth import routes
+from app.auth import routes  # noqa: E402, F401
