+++ app/templates/dashboard/dashboard.html
@@ -132,7 +132,7 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/feather-icons/4.9.0/feather.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datepicker/1.9.0/js/bootstrap-datepicker.min.js"></script>
 <script src="https://d3js.org/d3.v5.min.js"></script>
-<script type="text/javascript" src="{{ url_for('static', filename='dashboard.js') }}"></script>
+<script type="text/javascript" src="{{ url_for('dashboard.static', filename='dashboard.js') }}"></script>
 <script type="text/javascript">
   const load = () => {
     let startDate = $('#start-date').datepicker('getUTCDate');
