+++ app/templates/index.html
@@ -6,7 +6,7 @@
     <hr class="my-4">
     <p class="lead">first time users: see <a href="/#information">information section</a> below</p>
     <p class="lead">for reservations, see the <a href="/#gpu-calendar">calendar section</a> below</p>
-    <p class="lead">to view usage statistics, <a href="{{ url_for('home.dashboard') }}">click here</a></p>
+    <p class="lead">to view usage statistics, <a href="{{ url_for('dashboard.dashboard') }}">click here</a></p>
 </div>
 <div class="row">
     <div class="col-lg-12">
