+++ app/templates/layout.html
@@ -50,7 +50,7 @@
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav mr-auto">
         <li class="nav-item">
-          <a class="nav-link" href="{{ url_for('home.dashboard') }}">Dashboard</a>
+          <a class="nav-link" href="{{ url_for('dashboard.dashboard') }}">Dashboard</a>
         </li>
         {% if not current_user.is_anonymous %}
         <li class="nav-item">
